feat: Project RPG

+++ 02_CSharpAlgorithm/PersonalProject/ProjectRPG/Interfaces/ITargetable.cs
@@ -8,6 +8,6 @@ namespace ProjectRPG
 {
     internal interface ITargetable
     {
-        void Hit(int damage);
+        void Hit(float damage);
     }
 }
